services/bot/src/requirements.txt
openai
tiktoken
redis
discord
python-dotenv
ratelimit
asyncio